centralized http support test code
