+++ http-support-tests/src/test/scala/http/support/tests/server/Http4sServerSupportSpec.scala
@@ -1,0 +1,25 @@
+package http.support.tests.server
+
+import http.support.tests.UserCoding
+import typedapi.server._
+import typedapi.server.http4s._
+import cats.effect.IO
+import org.http4s.server.blaze.BlazeBuilder
+
+final class Http4sServerSupportSpec extends ServerSupportSpec[IO] {
+
+  import UserCoding._
+
+  val endpoints = deriveAll[IO](Api).from(path, segment, query, header, raw, get, put, putB, post, postB, delete)
+  val sm        = ServerManager(BlazeBuilder[IO], "localhost", 9000)
+  val server    = mount(sm, endpoints).unsafeRunSync()
+
+  "http4s implements TypedApi's server interface" >> {
+    tests(9000)
+
+    step {
+      server.shutdown.unsafeRunSync()
+    }
+  }
+
+}
